+++ app/controllers/AbsenceRequests.java
@@ -552,7 +552,8 @@ public class AbsenceRequests extends Controller {
     boolean approved = absenceRequestManager.approval(absenceRequest, user);
 
     if (approved) {
-      notificationManager.sendEmailToUser(Optional.fromNullable(absenceRequest), Optional.absent());
+      notificationManager.sendEmailToUser(Optional.fromNullable(absenceRequest), 
+          Optional.absent(), Optional.absent(), true);
 
       flash.success("Operazione conclusa correttamente");
     } else {
@@ -579,23 +580,20 @@ public class AbsenceRequests extends Controller {
         && user.hasRoles(Role.GROUP_MANAGER)) {
       // caso di approvazione da parte del responsabile di gruppo.
       absenceRequestManager.managerDisapproval(id, reason);
-      flash.error("Richiesta respinta");
-      render("@show", absenceRequest, user);
     }
     if (absenceRequest.administrativeApprovalRequired
         && absenceRequest.administrativeApproved == null && user.hasRoles(Role.PERSONNEL_ADMIN)) {
       // caso di approvazione da parte dell'amministratore del personale
       absenceRequestManager.personnelAdministratorDisapproval(id, reason);
-      flash.error("Richiesta respinta");
-      render("@show", absenceRequest, user);
     }
     if (absenceRequest.officeHeadApprovalRequired && absenceRequest.officeHeadApproved == null
         && user.hasRoles(Role.SEAT_SUPERVISOR)) {
       // caso di approvazione da parte del responsabile di sede
       absenceRequestManager.officeHeadDisapproval(id, reason);
-      flash.error("Richiesta respinta");
-      render("@show", absenceRequest, user);
     }
+    notificationManager.sendEmailToUser(Optional.fromNullable(absenceRequest), 
+        Optional.absent(), Optional.absent(), false);
+    flash.error("Richiesta respinta");
     render("@show", absenceRequest, user);
   }