+++ app/controllers/Configurations.java
@@ -254,13 +254,17 @@ public class Configurations extends Controller {
         .filter(conf -> conf.epasParam.category == EpasParam.EpasParamCategory.COMPETENCE_FLOWS)
         .collect(Collectors.toList());
     
+    final List<Configuration> informationFlows = configurations.stream()
+        .filter(conf -> conf.epasParam.category == EpasParam.EpasParamCategory.INFORMATION_FLOWS)
+        .collect(Collectors.toList());
+    
     // id relativo all'allegato di autorizzazione per l'attivazione dell'autocertificazione
     final Attachment autocert = office.attachments.stream()
         .filter(attachment -> attachment.type == AttachmentType.TR_AUTOCERTIFICATION).findFirst()
         .orElse(null);
 
     render(office, paramCategory, generals, yearlies, periodics, 
-        autocertifications, autocert, flows, competenceFlows);
+        autocertifications, autocert, flows, competenceFlows, informationFlows);
   }
 
   /**